TELEMED-AID-main/Backend/Medical Chatbot/requirements.txt
ctransformers
sentence-transformers
langchain
flask
flask-cors
pypdf
python-dotenv
langchain_community